Move dnd logic in QuestionDndContext component
